api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.58.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kagglehub==0.3.12
kiwisolver==1.4.8
matplotlib==3.10.5
mdurl==0.1.2
mkl-service
mpmath==1.3.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.1.2
orjson==3.10.16
passlib==1.7.4
pycparser==2.22
PyMySQL==1.1.2
pyparsing==3.2.3
PyQt6==6.7.1
python-dotenv==1.1.0
python-jose
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.0
Unidecode==1.4.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
psycopg2-binary>=2.9
cryptography
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29
aiomysql
pandas
pyarrow
